Added SimpleAgent.py
While this agent may need some refinement, it attempts to center the paddle below the ball, and does noticeably better than random, but not perfect. The main improvement it could use is attempting to find the ball when it is at the same row as some of the bricks that have disappeared.