Added a NEAT agent that has 3 inputs
Rather than using 520 inputs for pixels, this agent uses only three inputs: The position of the ball, the position of the paddle, and the speed of the paddle. This, (combined with preventing the best solution found from being removed due to stagnation) allows for it to get a better result than the original NEAT agent. This agent could perhaps be further optimized by lowering the number of outputs from 4 to three, since 0 and 1 are effectively identical outputs.